extract helper methods and components from active-file
