+++ lib/components/application.js
@@ -8,9 +8,14 @@ const openFile = remote.require(`${__dirname}/../open-file`);
 const { saveFile } = remote.require(`${__dirname}/../main`);
 const lastActiveFile = require('../last-active-file');
 const rejs = new Selfup();
+import saveOnClose from '../save-on-close';
+import loadSavedFiles from '../load-saved-files';
 
 const currentWindow = remote.getCurrentWindow();
-// console.log(lastActiveFile);
+
+saveOnClose();
+loadSavedFiles();
+
 class Application extends Component {
   constructor() {
     super();
