+++ lib/components/render-markdown-buttons.js
@@ -1,5 +1,5 @@
-import sendGist from './send-gist';
-import store from './store';
+import sendGist from '../send-gist';
+import store from '../store';
 import React from 'react';
 
 module.exports = (file) => {
