+++ lib/store.js
@@ -1,15 +1,13 @@
 import EventEmitter from 'events';
-import { readFileSync } from 'fs';
-import { remote, ipcRenderer } from 'electron';
-const { saveFile, saveFileOnClose, saveAllFiles, closeProcedure } = remote.require(`${__dirname}/main`);
-import saveOnClose from './save-on-close';
+import { ipcRenderer } from 'electron';
 
 const store = new EventEmitter();
-let files = JSON.parse(localStorage.getItem("files")) || [];
-store.emit('change');
-let idCounter = localStorage.getItem("idCounter") || 0;
 
-saveOnClose();
+let files = JSON.parse(localStorage.getItem("files")) || [];
+
+store.emit('change');
+
+let idCounter = localStorage.getItem("idCounter") || 0;
 
 store.all = () => files.concat([]);
 
@@ -111,7 +109,6 @@ store.saveGist = (fileName, gistUrl) => {
     }
     return Object.assign({}, file);
   });
-  console.log('made it through gist assignment');
   store.emit('change', files);
 };
 
@@ -122,19 +119,6 @@ ipcRenderer.on('file-opened', (event, file) => {
 ipcRenderer.on('file-saved', (event, fileName, id) => {
   store.updateFileName(fileName, id);
 });
-
-function loadSavedFiles() {
-  const savedFiles = JSON.parse(localStorage.getItem('files')) || [];
-  return savedFiles.map(file => {
-    try {
-      const newContent = readFileSync(file.fileName).toString();
-      return Object.assign(file, { content: newContent});
-    } catch (e) {
-      if (e.message.includes('ENOENT')) {return undefined;}
-      throw e;
-    }
-  }).filter(file => file);
-}
 
 store.on('change', (event) => {
   let jsonFiles = JSON.stringify(files);
