+++ lib/update-content.js
@@ -1,0 +1,3 @@
+import store from './store';
+
+module.exports = (file, event) => store.saveContent(file.id, event.target.value);
